requirements.txt
certifi==2019.9.11
chardet==3.0.4
gitdb2==2.0.6
GitPython==3.0.4
idna==2.8
langdetect==1.0.7
numpy==1.17.3
pandas==0.25.3
protobuf==3.10.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
rook==0.1.92
six==1.12.0
smmap2==2.0.5
urllib3==1.25.6
websocket-client==0.56.0